+++ src/provider/SignatureHelpProvider/SignatureFunc.ts
@@ -1,4 +1,4 @@
-import * as vscode from 'vscode';
+import type * as vscode from 'vscode';
 import type { CAhkFunc } from '../../AhkSymbol/CAhkFunc';
 import type { TAhkFileData } from '../../core/ProjectManager';
 import type { TAhkTokenLine } from '../../globalEnum';
@@ -6,48 +6,10 @@ import type { TBiFuncMsg } from '../../tools/Built-in/func.tools';
 import { getBuiltInFuncMD } from '../../tools/Built-in/func.tools';
 import { getDAWithPos } from '../../tools/DeepAnalysis/getDAWithPos';
 import { getFuncWithName } from '../../tools/DeepAnalysis/getFuncWithName';
-import { getLStr } from '../../tools/str/removeSpecialChar';
 import type { TLineFnCall } from '../Def/getFnRef';
 import { fnRefLStr } from '../Def/getFnRef';
 import { SignBiFn } from './SignatureFuncBuiltIn';
-
-function SignatureFuncSnipText2Sign(selectionRangeText: string): string {
-    const arr0: readonly string[] = selectionRangeText.split('\n');
-    const arr1: string[] = [];
-    for (const str of arr0) {
-        if ((/[\t ];/u).test(str)) {
-            arr1.push(str.slice(0, getLStr(str).length));
-        } else {
-            arr1.push(str);
-        }
-    }
-
-    return arr1.join('\n').trim();
-}
-
-function SignUserFn(fnSymbol: CAhkFunc, comma: number): vscode.SignatureHelp | null {
-    const {
-        paramMap,
-        selectionRangeText,
-    } = fnSymbol;
-
-    const parameters: vscode.ParameterInformation[] = [];
-    for (const { keyRawName } of paramMap.values()) {
-        parameters.push(new vscode.ParameterInformation(keyRawName));
-    }
-
-    const label: string = selectionRangeText.includes('\n')
-        ? SignatureFuncSnipText2Sign(selectionRangeText)
-        : selectionRangeText;
-    const SignatureInformation = new vscode.SignatureInformation(label);
-    SignatureInformation.parameters = parameters;
-
-    const Signature = new vscode.SignatureHelp();
-    Signature.signatures = [SignatureInformation];
-    Signature.activeSignature = 0;
-    Signature.activeParameter = comma;
-    return Signature;
-}
+import { SignUserFn } from './SignatureFuncUser';
 
 export function SignatureFunc(AhkFileData: TAhkFileData, position: vscode.Position): vscode.SignatureHelp | null {
     const { line, character } = position;
@@ -87,10 +49,12 @@ export function SignatureFunc(AhkFileData: TAhkFileData, position: vscode.Positi
                 if (brackets === 0) {
                     break; // not this func;
                 }
-            } else if (s === ',') {
+            } else if (s === ',' && brackets === 1) {
+                //
                 comma++;
             }
         }
+
         if (brackets !== 0) {
             const fnSymbol: CAhkFunc | null = getFuncWithName(upName);
             if (fnSymbol !== null) return SignUserFn(fnSymbol, comma);